Refactoring modules, add model for storing and parsing issue
